feat(routes)
Missing routes are added
