+++ src/server/routes/dashboard.route.js
@@ -3,7 +3,7 @@ const { Router } = require('express');
 const router = Router();
 
 router.get('/', (req, res) => {
-  res.render('public/dashboard/dashboard');
+  res.render('dashboard/dashboard');
 });
 
 module.exports = router;
