+++ src/server/routes/login.route.js
@@ -1,9 +1,9 @@
 const { Router } = require('express');
-const { showLogin, login } = require('../controllers/auth.controller');
+const { show, login } = require('../controllers/auth.controller');
 
 const router = Router();
 
-router.get('/', showLogin);
+router.get('/', show);
 
 router.post('/', login);
