+++ src/server/routes/swllog.route.js
@@ -1,0 +1,8 @@
+const { Router } = require('express');
+const { show } = require('../controllers/swllog.controller');
+
+const router = Router();
+
+router.get('/', show);
+
+module.exports = router;
