+++ src/server/server.js
@@ -35,7 +35,7 @@ db.connect();
 // Set view engine to HTML
 app.engine('html', require('ejs').renderFile);
 app.set('view engine', 'html');
-app.set('views', path.join(__dirname, '../client'));
+app.set('views', path.join(__dirname, '../client/public'));
 
 //static files
 app.use(express.static(path.join(__dirname, '../client/public')));
@@ -44,73 +44,10 @@ app.use(express.static(path.join(__dirname, '../client/public')));
 routes(app);
 
 app.get('/contestlog', (req, res) => { /* ... */ });
-app.get('/scanlog', (req, res) => { /* ... */ });
-app.get('/swllog', (req, res) => { /* ... */ });
-app.get('/hamlog', (req, res) => { /* ... */ });
-app.get('/signup', (req, res) => { /* ... */ });
-app.get('/mwlog', (req, res) => { /* ... */ });
-app.get('/vhflog', (req, res) => { /* ... */ });
-app.get('/auth', cors(), (req, res) => { /* ... */ });
 app.get('/logout', function (req, res) {
   req.session.destroy();
   res.send("logout success!");
 });
-
-/*app.get('/dashboard', (req, res) => {
-  // Check if the user is authenticated
-  logger.error('Error unico');
-  if (!req.user) {
-    // If not, redirect to the login page
-    res.redirect('/login');
-  } else {
-    // If yes, render the dashboard page
-    res.render('dashboard', { user: req.user });
-  }
-});*/
-
-app.get('/auth', function(req, res) {
-  res.json({ accessToken: accessToken });
-});
-
-// User schema for demonstration purposes
-const users = [];
-
-/*app.post('/auth', cors(), async (req, res) => {
-  const user = users.find(u => u.email === req.body.email);
-  if (user == null) {
-    return res.status(400).send('User not found');
-  }
-
-  try {
-    if (await bcrypt.compare(req.body.password, user.password)) {
-      const accessToken = jwt.sign(user, process.env.ACCESS_TOKEN_SECRET);
-      res.json({ accessToken: accessToken });
-    } else {
-      res.status(401).send('Invalid password');
-    }
-  } catch {
-    res.status(500).send('Error during authentication');
-  }
-});*/
-
-app.post('/signup', async (req, res) => {
-  try {
-    const hashedPassword = await bcrypt.hash(req.body.password, 10);
-    const user = { email: req.body.email, password: hashedPassword };
-    users.push(user);
-    res.status(201).send('User created');
-  } catch {
-    res.status(500).send('Error creating user');
-  }
-});
-
-/*app.post('/signin', (req, res) => {
-  res.header('Content-type', 'text/html');
-  // Insert Login Code Here
-  let username = req.body.username;
-  let password = req.body.password;
-  res.send(`Username: ${username} Password: ${password}`);
-});*/
 
 app.post('/reset-password', async (req, res) => {
   const email = req.body.email;
